+++ src/components/analyser/details/DetailsOverview.tsx
@@ -1,33 +1,36 @@
-// Based on Portfolio DetailsOverview
-
 import React from 'react';
 import {
   useTheme,
   makeStyles,
   createStyles,
   Theme,
+  Typography,
   Grid,
+  Paper,
+  ListItem,
+  List,
+  Accordion,
+  AccordionSummary,
+  AccordionDetails,
 } from '@material-ui/core';
+import ExpandMoreIcon from '@material-ui/icons/ExpandMore';
 import { useTranslation } from 'react-i18next';
 import { Stock, StockDetails } from '../../../analyser/APIClient';
-import TextOverText from '../TextOverText';
+import SectionDivider from './SectionDivider';
+import DetailsOverviewInfoBox from './DetailsOverviewInfoBox';
 
 // stylesheet for the Summary section
 const useStyles = makeStyles((theme: Theme) =>
   createStyles({
-    infoBox: {
-      outlineStyle: 'solid',
-      outlineColor: 'ba',
-      outlineWidth: '0.15rem',
-      margin: '1rem 0',
+    imageWrapper: {
+      width: '100%',
     },
-    vl: {
-      margin: '0.5rem 1rem',
-      height: '4rem',
-      alignSelf: 'center',
-      // TODO: use theme color
-      borderColor: 'grey',
+    paper: {
+      padding: theme.spacing(2),
+      margin: 'auto',
+      maxWidth: 500,
     },
+    intro: theme.typography.body1,
   })
 );
 
@@ -39,8 +42,7 @@ type DetailsOverviewProps = {
   stockDetails: StockDetails;
 };
 
-// returns the details page header
-const DetailsOverviewInfoBox: React.FC<DetailsOverviewProps> = ({
+const DetailsOverview: React.FC<DetailsOverviewProps> = ({
   stockOverview,
   stockDetails,
 }) => {
@@ -48,74 +50,68 @@ const DetailsOverviewInfoBox: React.FC<DetailsOverviewProps> = ({
   const theme = useTheme();
   const { t } = useTranslation();
 
-  // TODO: no hard coded colors
-  // TODO: update range to fit data from analytics
-  // convert a score to a color
-  function convertPerformanceToColor(num: number): string {
-    return num <= 0 ? '#D64745' : '#50E2A8';
-  }
-  function convertToPercent(num: number): string {
-    return `${(Math.round(num * 1000) / 10).toString()}%`;
-  }
-
   return (
-    <Grid
-      container
-      xs={12}
-      justify="space-around"
-      direction="row"
-      className={classes.infoBox}
-    >
-      <Grid item>
-        {/* country */}
-        <TextOverText
-          top={`${stockOverview.country}`}
-          bottom={t('stock.country')}
-          colorTop={theme.palette.primary.dark}
-          colorBottom={theme.palette.primary.dark}
+    <>
+      <SectionDivider section={t('analyser.details.summaryHeader')} />
+      <Grid container spacing={1}>
+        <DetailsOverviewInfoBox
+          stockOverview={stockOverview}
+          stockDetails={stockDetails}
         />
+        <Grid
+          container
+          alignItems="flex-start"
+          justify="space-evenly"
+          direction="row"
+          spacing={5}
+        >
+          <Grid item xs={3}>
+            <Paper className={classes.paper}>
+              <img
+                className={classes.imageWrapper}
+                alt="Company Pictrue"
+                src={stockOverview.picture.toString()}
+              />
+            </Paper>
+          </Grid>
+          <Grid item xs={9}>
+            <List>
+              <ListItem>
+                <Accordion>
+                  <AccordionSummary
+                    expandIcon={<ExpandMoreIcon />}
+                    aria-controls="panel2a-content"
+                    id="panel2a-header"
+                  >
+                    <Typography className={classes.intro}>
+                      {stockDetails.intro}
+                    </Typography>
+                  </AccordionSummary>
+                  <AccordionDetails>Test</AccordionDetails>
+                </Accordion>
+              </ListItem>
+              <ListItem>
+                <Typography className={classes.intro}>
+                  {t('stock.founded')}: {stockDetails.founded}
+                </Typography>
+              </ListItem>
+              <ListItem>
+                <Typography className={classes.intro}>
+                  {t('stock.fullTimeEmployees')}:{' '}
+                  {stockDetails.fullTimeEmployees}
+                </Typography>
+              </ListItem>
+              <ListItem>
+                <Typography className={classes.intro}>
+                  {t('stock.address')}: {stockDetails.address}
+                </Typography>
+              </ListItem>
+            </List>
+          </Grid>
+        </Grid>
       </Grid>
-
-      {/* devider 2 */}
-      <hr className={classes.vl} />
-      {/* box section 3 */}
-      <Grid item>
-        {/* currency */}
-        <TextOverText
-          top={`${stockOverview.currency}`}
-          bottom={t('stock.currency')}
-          colorTop={theme.palette.primary.light}
-          colorBottom={theme.palette.primary.dark}
-        />
-      </Grid>
-
-      {/* devider 2 */}
-      <hr className={classes.vl} />
-      {/* box section 3 */}
-      <Grid item>
-        {/* industry */}
-        <TextOverText
-          top={`${stockOverview.industry}`}
-          bottom={t('stock.industry')}
-          colorTop={theme.palette.primary.light}
-          colorBottom={theme.palette.primary.dark}
-        />
-      </Grid>
-
-      {/* devider 2 */}
-      <hr className={classes.vl} />
-      {/* box section 3 */}
-      <Grid item>
-        {/* dividend */}
-        <TextOverText
-          top={convertToPercent(stockOverview.div)}
-          bottom={t('stock.div')}
-          colorTop={convertPerformanceToColor(stockOverview.div)}
-          colorBottom={theme.palette.primary.dark}
-        />
-      </Grid>
-    </Grid>
+    </>
   );
 };
 
-export default DetailsOverviewInfoBox;
+export default DetailsOverview;
